op2: wifi: Enable 2.4GHz channel bonding
According to XDA Senior Member dreinulldrei, the WiFi configuration
file used in the OnePlus 3 and 3T is the default one provided by
Qualcomm. That’s not really an issue in and of itself, but the user
discovered that the default configuration disabled channel bonding in
the 2.4GHz frequency. 5GHz frequency networks have channel bonding
enabled (and if you have access to this frequency, then it is advised
you connect to it), but if your router only support the 2.4GHz
frequency then this trick may be useful for you.

Enabling channel bonding should theoretically double your wireless throughput (as long as your router supports channel bonding), as the channel width increases from 20MHz to 40MHz. This trick is quite easy to implement, as all one needs to do is modify one line in WCNSS_qcom_cfg.ini (located in /system/etc/wifi).#
https://www.xda-developers.com/how-to-double-the-wifi-speed-on-your-oneplus-3-3t/
Change-Id: I56e09bd173d156ed83e19d9071baa7d1df7a37f9